+++ database/migrations/Done/2020_12_11_105806_create_pricing_exception_lists_table.php
@@ -17,8 +17,8 @@ class CreatePricingExceptionListsTable extends Migration
             $table->id();
             $table->string('ProductNumber');
             $table->unsignedBigInteger('ProductId');
-            $table->integer('Price');
-            $table->integer('MAP');
+            $table->float('Price');
+            $table->float('MAP');
             $table->foreign('ProductId')->references('id')->on('products')->onDelete('cascade');
             $table->unsignedBigInteger('PriceExceptionId');
             $table->foreign('PriceExceptionId')->references('id')->on('pricing_exceptions')->onDelete('cascade');
